♻️ refactor(quality-phase2): split 739-line ComposeCard into focused modules
Pure extraction, no behavior change:
- ComposeCard.tsx (243) — outer shell + Header
- AgentCard.tsx (429) — per-agent card with activity rows and accordion
- StopSummary.tsx (69) — soft-stop notice + per-group summaries
- composeTokens.ts (34) — shared STATE/C/MONO tokens + getToolMeta
